refactor(go/adbc/driver/flightsql): factor out server-based tests
