+++ go/adbc/driver/flightsql/flightsql_adbc_test.go
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+// Tests that use the SQLite server example.
+
 package flightsql_test
 
 import (
@@ -49,10 +51,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -287,11 +287,9 @@ func TestADBCFlightSQL(t *testing.T) {
 
 	suite.Run(t, &DefaultDialOptionsTests{Quirks: q})
 	suite.Run(t, &HeaderTests{Quirks: q})
-	suite.Run(t, &AuthnTests{})
 	suite.Run(t, &OptionTests{Quirks: q})
 	suite.Run(t, &PartitionTests{Quirks: q})
 	suite.Run(t, &StatementTests{Quirks: q})
-	suite.Run(t, &TimeoutTestSuite{})
 	suite.Run(t, &TLSTests{Quirks: &FlightSQLQuirks{db: db}})
 	suite.Run(t, &ConnectionTests{})
 	suite.Run(t, &DomainSocketTests{db: db})
@@ -742,370 +740,6 @@ func (suite *HeaderTests) TestPrepared() {
 
 	suite.Contains(suite.Quirks.middle.recordedHeaders.Get("x-header-one"), "value 1")
 	suite.Contains(suite.Quirks.middle.recordedHeaders.Get("x-header-two"), "value 2")
-}
-
-type AuthnTests struct {
-	suite.Suite
-
-	s    flight.Server
-	db   adbc.Database
-	cnxn adbc.Connection
-}
-
-type AuthnTestServer struct {
-	flightsql.BaseServer
-}
-
-func (server *AuthnTestServer) GetFlightInfoStatement(ctx context.Context, cmd flightsql.StatementQuery, desc *flight.FlightDescriptor) (*flight.FlightInfo, error) {
-	md := metadata.MD{}
-	md.Set("authorization", "Bearer final")
-	if err := grpc.SendHeader(ctx, md); err != nil {
-		return nil, err
-	}
-	tkt, _ := flightsql.CreateStatementQueryTicket([]byte{})
-	info := &flight.FlightInfo{
-		FlightDescriptor: desc,
-		Endpoint: []*flight.FlightEndpoint{
-			{Ticket: &flight.Ticket{Ticket: tkt}},
-		},
-		TotalRecords: -1,
-		TotalBytes:   -1,
-	}
-	return info, nil
-}
-
-func (server *AuthnTestServer) DoGetStatement(ctx context.Context, tkt flightsql.StatementQueryTicket) (*arrow.Schema, <-chan flight.StreamChunk, error) {
-	sc := arrow.NewSchema([]arrow.Field{{Name: "a", Type: arrow.PrimitiveTypes.Int32, Nullable: true}}, nil)
-	rec, _, err := array.RecordFromJSON(memory.DefaultAllocator, sc, strings.NewReader(`[{"a": 5}]`))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	ch := make(chan flight.StreamChunk)
-	go func() {
-		defer close(ch)
-		ch <- flight.StreamChunk{
-			Data: rec,
-			Desc: nil,
-			Err:  nil,
-		}
-	}()
-	return sc, ch, nil
-}
-
-func authnTestUnary(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp interface{}, err error) {
-	md, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return nil, status.Error(codes.InvalidArgument, "Could not get metadata")
-	}
-	auth := md.Get("authorization")
-	if len(auth) == 0 {
-		return nil, status.Error(codes.Unauthenticated, "No token")
-	} else if auth[0] != "Bearer initial" {
-		return nil, status.Error(codes.Unauthenticated, "Invalid token for unary call: "+auth[0])
-	}
-
-	md.Set("authorization", "Bearer final")
-	ctx = metadata.NewOutgoingContext(ctx, md)
-	return handler(ctx, req)
-}
-
-func authnTestStream(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
-	md, ok := metadata.FromIncomingContext(ss.Context())
-	if !ok {
-		return status.Error(codes.InvalidArgument, "Could not get metadata")
-	}
-	auth := md.Get("authorization")
-	if len(auth) == 0 {
-		return status.Error(codes.Unauthenticated, "No token")
-	} else if auth[0] != "Bearer final" {
-		return status.Error(codes.Unauthenticated, "Invalid token for stream call: "+auth[0])
-	}
-
-	return handler(srv, ss)
-}
-
-func (suite *AuthnTests) SetupSuite() {
-	suite.s = flight.NewServerWithMiddleware([]flight.ServerMiddleware{
-		{Stream: authnTestStream, Unary: authnTestUnary},
-	})
-	suite.s.RegisterFlightService(flightsql.NewFlightServer(&AuthnTestServer{}))
-	suite.Require().NoError(suite.s.Init("localhost:0"))
-	suite.s.SetShutdownOnSignals(os.Interrupt, os.Kill)
-	go func() {
-		_ = suite.s.Serve()
-	}()
-
-	uri := "grpc+tcp://" + suite.s.Addr().String()
-	var err error
-	suite.db, err = (driver.Driver{}).NewDatabase(map[string]string{
-		"uri":                            uri,
-		driver.OptionAuthorizationHeader: "Bearer initial",
-	})
-	suite.Require().NoError(err)
-}
-
-func (suite *AuthnTests) SetupTest() {
-	var err error
-	suite.cnxn, err = suite.db.Open(context.Background())
-	suite.Require().NoError(err)
-}
-
-func (suite *AuthnTests) TearDownTest() {
-	suite.Require().NoError(suite.cnxn.Close())
-}
-
-func (suite *AuthnTests) TearDownSuite() {
-	suite.db = nil
-	suite.s.Shutdown()
-}
-
-func (suite *AuthnTests) TestBearerTokenUpdated() {
-	// apache/arrow-adbc#584: when setting the auth header directly, the client should use any updated token value from the server if given
-	stmt, err := suite.cnxn.NewStatement()
-	suite.Require().NoError(err)
-	defer stmt.Close()
-
-	suite.Require().NoError(stmt.SetSqlQuery("timeout"))
-	reader, _, err := stmt.ExecuteQuery(context.Background())
-	suite.NoError(err)
-	defer reader.Release()
-}
-
-type TimeoutTestServer struct {
-	flightsql.BaseServer
-}
-
-func (ts *TimeoutTestServer) DoGetStatement(ctx context.Context, tkt flightsql.StatementQueryTicket) (*arrow.Schema, <-chan flight.StreamChunk, error) {
-	if string(tkt.GetStatementHandle()) == "sleep and succeed" {
-		time.Sleep(1 * time.Second)
-		sc := arrow.NewSchema([]arrow.Field{{Name: "a", Type: arrow.PrimitiveTypes.Int32, Nullable: true}}, nil)
-		rec, _, err := array.RecordFromJSON(memory.DefaultAllocator, sc, strings.NewReader(`[{"a": 5}]`))
-		if err != nil {
-			return nil, nil, err
-		}
-
-		ch := make(chan flight.StreamChunk)
-		go func() {
-			defer close(ch)
-			ch <- flight.StreamChunk{
-				Data: rec,
-				Desc: nil,
-				Err:  nil,
-			}
-		}()
-		return sc, ch, nil
-	}
-
-	// wait till the context is cancelled
-	<-ctx.Done()
-	return nil, nil, ctx.Err()
-}
-
-func (ts *TimeoutTestServer) DoPutCommandStatementUpdate(ctx context.Context, cmd flightsql.StatementUpdate) (int64, error) {
-	if cmd.GetQuery() == "timeout" {
-		<-ctx.Done()
-		return -1, ctx.Err()
-	}
-	return -1, arrow.ErrNotImplemented
-}
-
-func (ts *TimeoutTestServer) GetFlightInfoStatement(ctx context.Context, cmd flightsql.StatementQuery, desc *flight.FlightDescriptor) (*flight.FlightInfo, error) {
-	switch cmd.GetQuery() {
-	case "timeout":
-		<-ctx.Done()
-	case "fetch":
-		tkt, _ := flightsql.CreateStatementQueryTicket([]byte("fetch"))
-		info := &flight.FlightInfo{
-			FlightDescriptor: desc,
-			Endpoint: []*flight.FlightEndpoint{
-				{Ticket: &flight.Ticket{Ticket: tkt}},
-			},
-			TotalRecords: -1,
-			TotalBytes:   -1,
-		}
-		return info, nil
-	case "notimeout":
-		time.Sleep(1 * time.Second)
-		tkt, _ := flightsql.CreateStatementQueryTicket([]byte("sleep and succeed"))
-		info := &flight.FlightInfo{
-			FlightDescriptor: desc,
-			Endpoint: []*flight.FlightEndpoint{
-				{Ticket: &flight.Ticket{Ticket: tkt}},
-			},
-			TotalRecords: -1,
-			TotalBytes:   -1,
-		}
-		return info, nil
-	}
-	return nil, ctx.Err()
-}
-
-func (ts *TimeoutTestServer) CreatePreparedStatement(ctx context.Context, req flightsql.ActionCreatePreparedStatementRequest) (result flightsql.ActionCreatePreparedStatementResult, err error) {
-	<-ctx.Done()
-	return result, ctx.Err()
-}
-
-type TimeoutTestSuite struct {
-	suite.Suite
-
-	s    flight.Server
-	db   adbc.Database
-	cnxn adbc.Connection
-}
-
-func (ts *TimeoutTestSuite) SetupSuite() {
-	ts.s = flight.NewServerWithMiddleware(nil)
-	ts.s.RegisterFlightService(flightsql.NewFlightServer(&TimeoutTestServer{}))
-	ts.Require().NoError(ts.s.Init("localhost:0"))
-	ts.s.SetShutdownOnSignals(os.Interrupt, os.Kill)
-	go func() {
-		_ = ts.s.Serve()
-	}()
-
-	uri := "grpc+tcp://" + ts.s.Addr().String()
-	var err error
-	ts.db, err = (driver.Driver{}).NewDatabase(map[string]string{
-		"uri": uri,
-	})
-	ts.Require().NoError(err)
-}
-
-func (ts *TimeoutTestSuite) SetupTest() {
-	var err error
-	ts.cnxn, err = ts.db.Open(context.Background())
-	ts.Require().NoError(err)
-}
-
-func (ts *TimeoutTestSuite) TearDownTest() {
-	ts.Require().NoError(ts.cnxn.Close())
-}
-
-func (ts *TimeoutTestSuite) TearDownSuite() {
-	ts.db = nil
-	ts.s.Shutdown()
-}
-
-func (ts *TimeoutTestSuite) TestInvalidValues() {
-	keys := []string{
-		"adbc.flight.sql.rpc.timeout_seconds.fetch",
-		"adbc.flight.sql.rpc.timeout_seconds.query",
-		"adbc.flight.sql.rpc.timeout_seconds.update",
-	}
-	values := []string{"1.1f", "asdf", "inf", "NaN", "-1"}
-
-	for _, k := range keys {
-		for _, v := range values {
-			ts.Run("key="+k+",val="+v, func() {
-				err := ts.cnxn.(adbc.PostInitOptions).SetOption(k, v)
-				var adbcErr adbc.Error
-				ts.ErrorAs(err, &adbcErr)
-				ts.Equal(adbc.StatusInvalidArgument, adbcErr.Code)
-				ts.ErrorContains(err, "invalid timeout option value")
-			})
-		}
-	}
-}
-
-func (ts *TimeoutTestSuite) TestRemoveTimeout() {
-	keys := []string{
-		"adbc.flight.sql.rpc.timeout_seconds.fetch",
-		"adbc.flight.sql.rpc.timeout_seconds.query",
-		"adbc.flight.sql.rpc.timeout_seconds.update",
-	}
-	for _, k := range keys {
-		ts.Run(k, func() {
-			ts.NoError(ts.cnxn.(adbc.PostInitOptions).SetOption(k, "1.0"))
-			ts.NoError(ts.cnxn.(adbc.PostInitOptions).SetOption(k, "0"))
-		})
-	}
-}
-
-func (ts *TimeoutTestSuite) TestDoActionTimeout() {
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.update", "0.1"))
-
-	stmt, err := ts.cnxn.NewStatement()
-	ts.Require().NoError(err)
-	defer stmt.Close()
-
-	ts.Require().NoError(stmt.SetSqlQuery("fetch"))
-	var adbcErr adbc.Error
-	ts.ErrorAs(stmt.Prepare(context.Background()), &adbcErr)
-	ts.Equal(adbc.StatusTimeout, adbcErr.Code, adbcErr.Error())
-}
-
-func (ts *TimeoutTestSuite) TestDoGetTimeout() {
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.fetch", "0.1"))
-
-	stmt, err := ts.cnxn.NewStatement()
-	ts.Require().NoError(err)
-	defer stmt.Close()
-
-	ts.Require().NoError(stmt.SetSqlQuery("fetch"))
-	var adbcErr adbc.Error
-	_, _, err = stmt.ExecuteQuery(context.Background())
-	ts.ErrorAs(err, &adbcErr)
-	ts.Equal(adbc.StatusTimeout, adbcErr.Code, adbcErr.Error())
-}
-
-func (ts *TimeoutTestSuite) TestDoPutTimeout() {
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.update", "1.1"))
-
-	stmt, err := ts.cnxn.NewStatement()
-	ts.Require().NoError(err)
-	defer stmt.Close()
-
-	ts.Require().NoError(stmt.SetSqlQuery("timeout"))
-	var adbcErr adbc.Error
-	_, err = stmt.ExecuteUpdate(context.Background())
-	ts.ErrorAs(err, &adbcErr)
-	ts.Equal(adbc.StatusTimeout, adbcErr.Code, adbcErr.Error())
-}
-
-func (ts *TimeoutTestSuite) TestGetFlightInfoTimeout() {
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.query", "0.1"))
-
-	stmt, err := ts.cnxn.NewStatement()
-	ts.Require().NoError(err)
-	defer stmt.Close()
-
-	ts.Require().NoError(stmt.SetSqlQuery("timeout"))
-	var adbcErr adbc.Error
-	_, _, err = stmt.ExecuteQuery(context.Background())
-	ts.ErrorAs(err, &adbcErr)
-	ts.NotEqual(adbc.StatusNotImplemented, adbcErr.Code, adbcErr.Error())
-}
-
-func (ts *TimeoutTestSuite) TestDontTimeout() {
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.fetch", "2.0"))
-	ts.NoError(ts.cnxn.(adbc.PostInitOptions).
-		SetOption("adbc.flight.sql.rpc.timeout_seconds.query", "2.0"))
-
-	stmt, err := ts.cnxn.NewStatement()
-	ts.Require().NoError(err)
-	defer stmt.Close()
-
-	ts.Require().NoError(stmt.SetSqlQuery("notimeout"))
-	// GetFlightInfo will sleep for one second and DoGet will also
-	// sleep for one second. But our timeout is 2 seconds, which is
-	// per-operation. So we shouldn't time out and all should succeed.
-	rr, _, err := stmt.ExecuteQuery(context.Background())
-	ts.Require().NoError(err)
-	defer rr.Release()
-
-	ts.True(rr.Next())
-	rec := rr.Record()
-
-	sc := arrow.NewSchema([]arrow.Field{{Name: "a", Type: arrow.PrimitiveTypes.Int32, Nullable: true}}, nil)
-	expected, _, err := array.RecordFromJSON(memory.DefaultAllocator, sc, strings.NewReader(`[{"a": 5}]`))
-	ts.Require().NoError(err)
-	defer expected.Release()
-	ts.Truef(array.RecordEqual(rec, expected), "expected: %s\nactual: %s", expected, rec)
 }
 
 type TLSTests struct {
